i'm stupid
- after following various resources and guidance from a friend, my first AJAX request has been completed.

- Switch statements updated for error hints and redirecting to the same page or account depending on error got.

- I'm not sure why it took so many attempts when I kind of had it correct in the first place? but this is my understanding of it

---

THe reason it didn't work at first using a simple echo of the error was because comments were being put into the output stream. This meant i couldn't define a singular string for my switch statement.

Then I went on a wild goose chase because I must've been wrong, but in matter of fact it was genuinely quite simple  to setup.

I'n future i'd like to change how i handle errors properly, i  need to do more research on it but for now this works as gest as I think it does.